removed unneccessary props for GeneralInformation, changed typization for useWatch
