--- src/pages/deliveryForm/steps/generalInformation2/generalInformation.scss
@@ -1,43 +1,0 @@
-.general-information-form {
-	&__stepper {
-		margin-bottom: 40px;
-		@media (max-width: 48em) {
-			margin-bottom: 20px;
-		}
-	}
-	&__form {
-	}
-	&__row {
-		display: grid;
-	}
-	&__row--1 {
-		grid-template-columns: 1fr 1fr 2fr;
-		gap: 10px;
-		margin-bottom: 14px;
-		@media (max-width: 48em) {
-			grid-template-columns: 1fr;
-		}
-	}
-	&__row--2 {
-		&:not(:last-child) {
-			margin-bottom: 20px;
-		}
-	}
-	&__row--3 {
-	}
-	&__row--4 {
-	}
-	&__row--5 {
-	}
-	&__row--6 {
-		justify-items: center;
-		padding-top: 40px;
-	}
-	&__promocode {
-		max-width: 220px;
-		padding: 0 0 40px;
-		box-shadow: 0 0 0 100vmax var(--mainGrayColor);
-		clip-path: inset(0 -100vmax);
-		background-color: var(--mainGrayColor);
-	}
-}
